Windows: Handle environment variable inheritance uniformly in Mesos.
This review is a planned regression on MESOS-6816. In other parts of the
codebase, we copy all environment variables from the system before
launching a process, overwriting user specified environment variables.

This is not correct behavior, but a half-measure that is necessary,
because Windows does not inherit environment variables by default. In
this commit, we make this behavior uniform across all places where we
are creating a process, because it is better for behavior to be
consistent before we get around to fixing it. We do have concrete plans
to come back and resolve this issue (and allow users to specify certain
environment variables).

Review: https://reviews.apache.org/r/56702/