Libprocess: Dealt with some MSVC compiler warnings.
This commit explicitly qualifies some implicit number casting.

The only additional CHECK added by this commit is in the `FileEncoder`,
which casts an unsigned size (size_t) to a signed value (off_t).
All usages of `FileEncoder` are expected to retrieve the size value
from the `off_t st_size` field of `struct stat`.  This means that
it is a programmer error to construct `FileEncoder` with a size
greater than the numeric maximum of `off_t`.

Review: https://reviews.apache.org/r/56592/