Stout: Explicitly marked protobuf file as using v2 syntax.
`protoc` will complain about `.proto` files that do not explicitly
specify the version of PB syntax they're using. On Windows, this
registers as a build warning.

To ameliorate this, we explicitly specify the syntax in all `.proto`
files.

NOTE: The generated protobuf does not need to be updated because
this commit does not change the content of the protobuf.  This
change only explicitly marks the syntax of the file.

Review: https://reviews.apache.org/r/56594/