CMake: Add `build_config.hpp.in` for `BUILD_TIME` variables.
Commit c7fc1377b introduced a bug that prevented any incremental
recompilation.  By defining the `BUILD_TIME` in `MESOS_CPPFLAGS`,
every build was seen as having a root dependency (the flags)
change, causing a rebuild of every single source in Mesos
(especially bad on ccache-less systems).

This patch introduces a CMake `configure_file` directive which
takes in `build_config.hpp.in` and emits `build_config.hpp` with the
`BUILD_TIME`, `BUILD_DATE`, and `BUILD_USER` variables defined.
As this is a CMake-only file, we guard the new header with a static
definition `-DUSE_CMAKE_BUILD_CONFIG`.  The result is that the date,
time, and user are set at the point of configuration (invocation of
CMake) instead of at build, thus allowing for incremental rebuilds.

Review: https://reviews.apache.org/r/57052/