CMake: Refactored FindApr.cmake and FindSvn.cmake.
The logic to find APR and SVN headers/libraries are effectively the
same, except for where those headers/libraries might be found.

This commit refactors the two files into a helper that takes several
lists as arguments (lists are passed by reference in CMake) and
outputs two lists of directories.

This also removes the logic for finding `APR-util`, which is not
used in this project.

Review: https://reviews.apache.org/r/57260