Renamed `activeRoles` to `roles` in the master.
The notion of "active" currently refers to an entity's state in regards
to the flow of resources. For example, a framework is said to be
"active" if it is receiving resources, and an agent is said to be
"active" if it is sending resources. `activeRoles` however does not
follow this notion of "active", since it simply consists of roles with
at least one __registered__ framework, rather than __active__.

Instead, it is a list of roles that the master is keeping track of.
We therefore opt to drop the "active" portion of this list of roles.

Review: https://reviews.apache.org/r/57329