Fixed a bug in master and agent handler authorization logic.
This patch fixes a bug where endpoint handlers would not
correctly handle the case in which authorization is enabled
when authentication is disabled. In this case, the handlers
would send a default-constructed `authorization::Subject` to
the authorizer, leading to an empty-string principal being
evaluated as the subject.

This patch updates the handlers to correctly send `NONE` as
the subject in this case.

Review: https://reviews.apache.org/r/57054/