Fixed crash with tasks that use very small resource values.
When parsing resources, inputs such as "cpus:0" are considered "empty"
and are omitted from the resulting `Resources` object. However, a very
small resource value like "cpus:0.00001" was considered non-empty,
despite the fact that the numerical behavior for resources means that
these two values compare as equal. This inconsistency resulted in a
`CHECK` failure in the sorter when a task that used a tiny amount of
resources reached a terminal state.

This commit fixes the resource parsing logic to treat such very small
resource values as empty resources, which resolves the inconsistency
above. The resulting behavior might be a bit surprising to users (since
the task will launch with zero resources instead of a very small
resource value), but improving this is left as future work (MESOS-1807).

Review: https://reviews.apache.org/r/57752