Replaced an instance of `/dev/null` with `os::DEV_NULL`.
`os::DEV_NULL` is a constant defined in stout that points to `/dev/null`
on Posix and `NUL` on Windows.  The two locations behave similarly
on their respective platforms.

This change future-proofs this code on Windows (only fails at runtime)
in case the related features are added to the Windows build.

Review: https://reviews.apache.org/r/57480