Changed DRFSorter's representation of inactive clients.
DRFSorter previously removed inactive clients from the `clients`
collection, and then re-added clients when they were reactivated. This
resulted in resetting the allocation count for the client, which is
unfortunate. This scheme would also be more difficult to adapt to
hierarchical sorting.

This commit changes DRFSorter to continue to store inactive clients in
the `clients`; inactive clients are indicated by a new field in the
`Client` struct, and are omitted from the return value of
`DRFSorter::sort`.

Review: https://reviews.apache.org/r/57564