Windows: Added Glog patch for enabling stack trace.
This relies on an updated `glog-da816ea70.tar.gz` tarball
in the Mesos 3rdparty repository, which is itself generated from
the `da816ea70` commit on github.com/google/glog.

Additionally, this applies a patch that includes stacktrace support
to glog on Windows.  The code in this patch is included in an upstream
pull request:
https://github.com/google/glog/pull/168

If and when that pull request gets merged, we will update the glog
tarball and remove the need for the extra patch file.

Review: https://reviews.apache.org/r/57490/