Windows: Updated libprocess CMake setup for Glog patch.
The additional Windows library `Dbghelp` must be linked for the
in addition to glog, in order to enable stack traces.

NOTE: CMake's dependency graph does not pull in the `Dbghelp` library
automatically as the glog dependency is added into Mesos as an
"external" project.  If we were to, instead, add glog's CMake files
to the build system directly (such as, as a git submodule), glog's
targets would be inherited.

Review: https://reviews.apache.org/r/57491/