Check task user before allowing a task to be launched on the agent.
Added support for action `run_tasks` on the agent's flag `acl`. Based on
the ACL configured for `run_tasks`, a task to be launched on the agent
can be (dis)allowed to launch on the agent.
If a task or task group cannot be launched due to failed authorization,
a `TASK_ERROR` status update shall be sent with a reason code of
`REASON_TASK_UNAUTHORIZED` or `REASON_TASK_GROUP_UNAUTHORIZED` as
applicable.

Note that in case of a task group, all tasks fail if any of the tasks
within the task group encounter the authorization error.

Review: https://reviews.apache.org/r/55887/