Modified CNI isolator to treat "__MESOST_TEST__*" as a test network.
Currently the CNI isolator identifies the "__MESOS_TEST__" as a test
network. However, there are cases where we might need multiple test
networks to test against. To support this use case we now do a
substring match on the network name and treat any network with name of
the form "*__MESOS_TEST__*" as a test network.

Review: https://reviews.apache.org/r/57627/