Fixed `post-reviews.py` to work on Windows.
Windows cannot find `rbt` because the command name is `rbt.cmd`;
the extension is required.

Because Windows (PowerShell and cmd) does not support shebangs,
the usage of this script on Windows is:

    python support\post-reviews.py

Windows does not have the `cat` command, so replace its usage
with `git rev-parse --verify <branch>`.

Using `os.path.join` for URLs on Windows does not work,
as the Windows path separator is `\` instead of `/`.
This broke updating previously submitted reviews.
We fix it by using `urlparse.urljoin` for the joining the
base URL with rest of the URL.

Note that colors do not always work on Windows, but can work,
so we leave them enabled.

Review: https://reviews.apache.org/r/57857/