Fixed UpdateFrameworkMessage handling backwards incompatibility.
The commit 4912f341dce12f10b43fe82db20086032391e95a introduced a new
`framework_info` field in the `UpdateFrameworkMessage` sent from the
master to the agent. Unfortunately, the agent code unconditionally
uses the new field, which results in a default-intialized
`FrameworkInfo` when a new agent is receiving the message from an
old master. This leads to an agent abort when it subsequently tries
to launch a new task for the updated framework.

This fixes the issue by ignoring the field if unset.

Review: https://reviews.apache.org/r/57836/