Changed the namespaces of the secret generators.
This patch updates the namespaces of the secret generators.
The `SecretGenerator` is placed into `mesos::`, consistent
with other classes declared in files in the same directory.
The implementations in `src/authentication/` do not exhibit
a consistent pattern, so the `JWTSecretGenerator` was placed
in a namespace indicative of its directory structure:
`mesos::authentication::executor`.

Review: https://reviews.apache.org/r/57840/