Turned off executor authentication in the ExecutorHttpApiTests.
This patch turns off executor authentication in the
ExecutorHttpApiTests, since authentication is not related
to the functionality they are meant to test. Since executor
tokens depend on the FrameworkID/ExecutorID/ContainerID of
the executor, it would be difficult to enable it in these
tests.

Review: https://reviews.apache.org/r/57750/