Removed `allocations` map from DRFSorter.
The sorter previously managed three collections: a set of active
clients, a map from client names to allocations, and a map from client
names to weights. Since the set previously contained only active
clients, each client's allocation needed to be stored separately, since
the sorter needed to remember the allocation made to inactive clients.

Now that the set of clients includes both active and inactive clients,
we can dispense with the additional `allocations` map and store a
client's allocation as a nested struct in the `Client` struct. This
avoids the need to keep the two collections synchronized; the logic for
manipulating allocations is also more properly written as a member
function of the Client::Allocation struct, rather than inline in a
member function of DRFSorter.

Note that this change significantly increases the size of the `Client`
struct. Since we copy clients by value in several places (e.g., when
erasing and re-inserting a `Client` into `std::set`), this change
regresses sorter performance. This will be fixed in the next review in
this chain: we'll move from `std::set<Client>` to `std::set<Client*>`,
which will avoid the need to copy the `Client` on updates.

Review: https://reviews.apache.org/r/58110