Combined containerizer interface's launch methods.
When nested container support was added, we added a separate `launch`
path in the containerizer because nested containers do not need
an explicit TaskInfo/ExecutorInfo.  Nested containers basically
only need the CommandInfo and ContainerInfo.

This commit combines the two launch methods by replacing most of the
"Infos" (Task, Executor, Command, Container) with a `ContainerConfig`
argument, which may contain multiple combinations of the "Infos".

The goal is to support three launch paths for containers:
  1) When the `ContainerConfig` contains a TaskInfo/ExecutorInfo,
     launch a task or executor.
  2) When the `ContainerID` has a parent, launch a nested container.
  3) (Not implemented yet) When there is no TaskInfo/ExecutorInfo or
     parent container, launch a standalone container.

There are two other notable changes to the interface:
  * The `SlaveID` field has been removed entirely.  The code that
    requires this (in the fetcher and Docker containerizer) will be
    addressed in a separate commit.
  * The `checkpoint` bool has been replaced by an Option<string>,
    which contains the path that should be used for checkpointing.
    This path includes the filename.
    This is also one of the reasons why `SlaveID` was an argument.

Review: https://reviews.apache.org/r/58899