Fixed linting script when run with no arguments.
The python linter needs to be run in a virtualenv due to how
it requires multiple modules that may or may not be installed
(or the correct version) in the host environment.  We originally
re-made this virtualenv whenever some specific files were
changed (such as the pip-requirements.txt, which describes
the virtualenv's dependencies).

However, this logic was missing the case where we ran
support/mesos-style.py with no arguments.  In this case, the
entire codebase is linted.  Therefore, the virtualenv should
also be rebuilt in this case.