Windows: Fixed Base64Test.EncodeURLSafe.
C++ encodes string literals in the compiling platform's encoding
of choice, which means UTF8 for Posix, and ANSI for Windows.

This has implications for this particular test, as the string literal
"~~~\u00ff\u00ff\u00ff\u00ff" is translated into different bytes:
  Posix:   { 126, 126, 126, 195, 191, 195, 191, 195, 191, 195, 191 }
  Windows: { 126, 126, 126,      255,      255,      255,      255 }

Prepending `u8` to the string literal tells the compiler to encode
the string as UTF8.  This does not expose any underlying bug(s)
on Windows because the test is only failing due to an incorrect input.

Review: https://reviews.apache.org/r/58430/