Prevent agent from crashing when the ID is too long.
When launching a task with a very long name mkdir will fail and crash
the agent, making it quick and easy to remotely crash any agent(s).
This changes makes sure that we never exceed the default file name
limit of most of the file systems out there. If the limit is exceed a
TASK_ERROR is instead sent back to the client.

Review: https://reviews.apache.org/r/58317/

*** Modified for 1.2.x. ***