Fixed a potential deadlock in the 'CombinedAuthenticator'.
This patch fixes a bug in `CombinedAuthenticator.authenticate()`.
Previously, the function registered a callback on the results of
`Authenticator::authenticate()` calls, which would be executed in
the context of the individual authenticators' processes. Since this
callback captured a copy of the `Owned<Authenticator>` referring to
that authenticator itself, it was possible that during teardown the
authenticator could be destroyed from its own context, leading to
a deadlock.

This patch eliminates the capture of a reference to the
authenticator, instead simply capturing the authenticator's scheme.

Review: https://reviews.apache.org/r/58464/