Windows: Updated ZooKeeper to use CMake.
This unblocks us from building exclusively with VS 2017. The previous
patch to ZooKeeper only added VS 2015 support. This patch replaces it
with a CMake build system that will generate whichever solution we need
for Windows (and can replace the Autotools system on Linux).

We're updating to 3.5.2-alpha as the existing 3.5.1 rebundle was a
source tarball, and so missing the necessary generated files. The most
currently used version was based off a random commit. 3.5.2-alpha is the
latest 3.5.x release of ZooKeeper (3.5.x itself is alpha, 3.5.2 is
semi-stable, in comparison to 3.5.3 which is in RC).

Review: https://reviews.apache.org/r/58448/