Updated the Overwrite docker provisioner test for MESOS-7280.
Created a non-empty file 'abc' under the agent's work directory, which
is linked by the symlink '/xyz' in the 2nd layer of the testing image.
This file should not be overwritten by the empty file '/xyz' in the 3rd
layer. Please see the following link for more details:
  https://hub.docker.com/r/chhsiao/overwrite/

The testing image is updated in a way that it is compatible to the
previous Overwrite test.

Review: https://reviews.apache.org/r/58640/