Changed naming of Docker containers to the pre-0.23 scheme.
For the most part, the choice of naming scheme is cosmetic.
All versions of the Docker containerizer will consider containers
starting with "mesos-" as containers started by Mesos.
The Docker containerizer does not consider the `SlaveID` at all
when recovering.

This naming change also tweaks the recovery logic to store the
name of the recovered container, so that future calls to the Docker
CLI use the actual name of the container, rather than a name generated
based on the ContainerID.

Review: https://reviews.apache.org/r/58905