Avoid a corruption while rescinding offers.
When a `DESTROY` is processed, we rescind all pending offers to which
the destroyed persistent volumes have been offered. As soon as we
rescind an offer, the `Offer` object is freed; and hence, we should
move on to the next offer (even though there could be multiple volumes
being destroyed in the same `ACCEPT` call).

Review: https://reviews.apache.org/r/58485/