Updated all tests that use Containerizer::launch(...).
This sweeps through the unit tests and translates all existing
tests that use the two variants of Containerizer::launch(...)
(i.e. nested and non-nested) and translates the arguments to the
new interface.  Some of the  fetcher interface changes are also
addressed in this sweep.

Review: https://reviews.apache.org/r/59074