Fixed provisioner recover blockage by non-existing rootfses dir.
In provisioner recover, when listing the container rootfses, it is
possible that the 'rootfses' dir does not exist. Because a possible
race between the provisioner destroy and the agent restart. For
instance, while the provisioner is destroying the container dir the
agent restarts. Due to os::rmdir() is recursive by traversing the
FTS tree, it is possible that 'rootfses' dir is removed but the
others (e.g., scratch dir) are not.

Currently, we are returning an error if the 'rootfses' dir does not
exist, which blocks the agent from recovery. We should skip it if
'rootfses' does not exist.

Review: https://reviews.apache.org/r/59061/