Cleaned up pattern matching in Resource operator==.
The usual pattern for checking optional fields here is to check
whether the field is is either set for both sides or unset for both.
Subsequentially we check if possible values are equal, returning
'false' if they are not.

The code changed here did not follow this pattern, but instead assumed
it was the last check to run. This made it possible that a careless
addition after these checks would never be called.

Review: https://reviews.apache.org/r/58136/