Added workaround to unbreak the build with glibc <= 2.12.
In old versions of glibc, the <sys/sysmacros.h> header must be enclosed
in an `extern "C" { ... }` block. This prevents Mesos from linking on
CentOS 6.9 and likely other platforms from the same timeframe.

In more recent versions of glibc, the `extern "C" { ... }` linkage
declaration is not necessary (because the header file specifies this
itself). In this situation, there will now be two linkage declarations,
but that is fine: the inner-most linkage declaration takes precedence.

Review: https://reviews.apache.org/r/59154