Windows: Updated `support/windows-build.bat`.
This helper script now understands the environment variable
`CMAKE_GENERATOR` so that users can change the generator;
the default is still Visual Studio 14.

Direct usage of `msbuild` was replaced with `cmake --build`.

We now build `mesos-tests` in the same pattern as `stout-tests` and
`libprocess-tests`. If admin privileges are missing, we skip running the
tests, but still proceed to build the general target.

We set the CMake toolset instead of an environment variable for
`PreferredToolArchitecture`.

Review: https://reviews.apache.org/r/59116/